Split up RenameOperation
