+++ tests/Components/RenameOperationTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace PhpMyAdmin\SqlParser\Tests\Components;
 
-use PhpMyAdmin\SqlParser\Components\RenameOperation;
+use PhpMyAdmin\SqlParser\Components\Lists\RenameOperations;
 use PhpMyAdmin\SqlParser\Parser;
 use PhpMyAdmin\SqlParser\Tests\TestCase;
 
@@ -12,7 +12,7 @@ class RenameOperationTest extends TestCase
 {
     public function testBuildAll(): void
     {
-        $component = RenameOperation::parse(new Parser(), $this->getTokensList('a TO b, c TO d'));
-        $this->assertEquals(RenameOperation::buildAll($component), 'a TO b, c TO d');
+        $component = RenameOperations::parse(new Parser(), $this->getTokensList('a TO b, c TO d'));
+        $this->assertEquals(RenameOperations::buildAll($component), 'a TO b, c TO d');
     }
 }
